# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Personal AI agent management API — multi-agent routing, LLM providers, tools, approvals, and more.",
        "title": "Denkeeper API",
        "contact": {
            "name": "Denkeeper",
            "url": "https://github.com/Temikus/denkeeper"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/channels": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all configured channels with their agent bindings, adapter associations, and active adapter keys.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "channels"
                ],
                "summary": "List all channels",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_api.channelResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a named channel bound to an agent with optional adapter bindings, delivery mode, and session mode. Persists the channel to the TOML config file.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "channels"
                ],
                "summary": "Create a new channel",
                "parameters": [
                    {
                        "description": "Channel configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.channelCreateInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_api.channelResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error (missing name/agent, invalid delivery or session_mode)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Referenced agent not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Channel already exists or channels not configured",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Config persistence not available",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/channels/{name}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single channel's configuration including agent binding, adapters, delivery/session mode, conversation ID, and active adapter keys.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "channels"
                ],
                "summary": "Get a channel by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.channelResponse"
                        }
                    },
                    "404": {
                        "description": "Channel not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes a channel from the dispatcher and the TOML config file. Implicit channels cannot be deleted.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "channels"
                ],
                "summary": "Delete a channel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Channel deleted"
                    },
                    "400": {
                        "description": "Channel is implicit",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Channel not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Config persistence not available",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Partially updates a channel's configuration (agent, adapters, delivery, session_mode). Only provided fields are applied; omitted fields retain their current values. Implicit channels cannot be modified. Persists changes to the TOML config file.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "channels"
                ],
                "summary": "Update an existing channel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update (all optional except at least one)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.channelCreateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.channelResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON, invalid delivery/session_mode, or channel is implicit",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Channel or referenced agent not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Config persistence not available",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api.channelCreateInput": {
            "type": "object",
            "properties": {
                "adapters": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "agent": {
                    "type": "string"
                },
                "delivery": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "session_mode": {
                    "type": "string"
                }
            }
        },
        "internal_api.channelResponse": {
            "type": "object",
            "properties": {
                "active_adapter_keys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "adapters": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "agent": {
                    "type": "string"
                },
                "conversation_id": {
                    "type": "string"
                },
                "delivery": {
                    "type": "string"
                },
                "implicit": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "session_mode": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "API key prefixed with \"Bearer \"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}